update findOne prehook and create can-access-calendar-event.provider
