+++ packages/twenty-server/src/modules/calendar/query-hooks/calendar-query-hook.module.ts
@@ -7,6 +7,7 @@ import { CalendarChannelEventAssociationObjectMetadata } from 'src/modules/calen
 import { CalendarChannelObjectMetadata } from 'src/modules/calendar/standard-objects/calendar-channel.object-metadata';
 import { CalendarEventFindManyPreQueryHook } from 'src/modules/calendar/query-hooks/calendar-event/calendar-event-find-many.pre-query.hook';
 import { CalendarEventFindOnePreQueryHook } from 'src/modules/calendar/query-hooks/calendar-event/calendar-event-find-one.pre-query-hook';
+import { CanAccessCalendarEventProvider } from 'src/modules/calendar/query-hooks/calendar-event/providers/can-access-calendar-event.provider';
 
 @Module({
   imports: [
@@ -18,6 +19,7 @@ import { CalendarEventFindOnePreQueryHook } from 'src/modules/calendar/query-hoo
     ]),
   ],
   providers: [
+    CanAccessCalendarEventProvider,
     {
       provide: CalendarEventFindOnePreQueryHook.name,
       useClass: CalendarEventFindOnePreQueryHook,
